requirements.txt
requests==2.31.0
python-dotenv==1.0.0
chromadb==0.4.24
cohere==5.0.3
langchain==0.1.10
pandas==2.1.4
numpy==1.26.3
openai==1.3.9
